Fix parallel handling of data in repeated trajectory runs for 'rotate' and 'average' (#1113)
* Add test for handling generated data set vs read-in dataset

* Test that a generated set can be used as a read-in set once processing
is complete.

* Add broadcast function

* Ensure set is properly broadcast if it was previously generated but we
are using it in a new trajectory processing run

* Enable parallel dataset handling test

* Add broadcast to the coords frames data set. Make sure we use long int when broadcasting dataset size.

* Add data broadcast to frames coords set

* No need for a broadcast, all data is on the disk

* Add broadcast

* Add broadcast for string set

* Add Sync and Bcast to tensor dataset

* Add bcast

* Add broadcast for vector data

* Add bcast

* Ensure 'average' broadcasts the final average coords set to non-master
ranks for future use. Addresses #1096.

* Version 6.29.7. Revision bump for fixed broadcasting of data created
during trajectory processing in a subsequent trajectory processing run
for rotate and average commands.

---------

+++ src/DataSet_Coords_CRD.cpp
@@ -183,8 +183,8 @@ int DataSet_Coords_CRD::Sync(size_t total, std::vector<int> const& rank_frames,
 int DataSet_Coords_CRD::Bcast(Parallel::Comm const& commIn) {
   if (commIn.Size() == 1) return 0;
   // Assume all data is currently on the master process.
-  int totalSize = Size();
-  int err = commIn.MasterBcast( &totalSize, 1, MPI_INT );
+  long int totalSize = Size();
+  int err = commIn.MasterBcast( &totalSize, 1, MPI_LONG );
   if (!commIn.Master()) {
     //rprintf("DEBUG: Resizing array to %i\n", totalSize);
     frames_.Resize( totalSize );
